tools/mcp/memory/requirements.txt
mcp>=1.0.0
fastembed>=0.2.0
faiss-cpu>=1.7.0
numpy
